requirements.txt
torch --extra-index-url https://download.pytorch.org/whl/cu113
numpy
torchvision
albumentations
tqdm
matplotlib